requirements.txt
Flask==3.0.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
openpyxl==3.1.2
Werkzeug==3.0.1
scikit-learn==1.3.2
statsmodels==0.14.0
google-generativeai==0.3.2
